fix: Handle all incident capture features in reporter app
